(update): powersync is working now.

+++ lib/features/products/domain/usecases/upload_products_from_excel_usecase.dart
@@ -41,8 +41,8 @@ class UploadProductsFromExcelUseCase {
               id: 'someidchange later',
               name: name,
               categoryId: 'catid',
-              listPrice: double.parse(price!),
-              standardPrice: double.parse(cost!),
+              listPrice: int.parse(price!),
+              standardPrice: int.parse(cost!),
               defaultCode: code));
         }
       }
